requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.30.0
